Add `VideoEdit` `Volume` action
